Update README.md
Attrition refers to the percentage of employees who leave a company during a specific time, this includes: resignation, termination, firing and retirement. Attrition rate of a company should be as low as possible as it may lead to impact the company's working capability as well as males the hiring process more challenging.
From the dashboard the following insights can be drawn:
The attrition rate of the company is 16%, which is way too high for an organization to run efficiently.
A large number of employees left the company within 4 years of joining.
Working overtime, frequent travels and job satisfaction are the major factors for driving the attrition rate.
The employees within the range of 25 to 35 years are the most likely to leave the company.
There should be a proper work life balance in an employee's life so as to retain them.